+++ link-shortener/src/assets/styles.css
@@ -1,5 +1,3 @@
-function stylesheet() {
-	return `
 body {
 	background-color: #fff;
 	font-family: sans-serif;
@@ -27,7 +25,3 @@ body {
 	max-width: 100%;
 	padding: 0 15px;
 }
-`;
-};
-
-export default stylesheet;
